renamed all the types to reduce chance of name collisions

+++ client/src/graphql/mutations/addShipment.ts
@@ -1,7 +1,7 @@
 import { gql } from "@apollo/client"
   
 const ADD_SHIPMENT = gql`
-  mutation addShipment($pickupLocation: [Int]!, $dropoffLocation: [Int]!, $description: String!) {
+  mutation addShipment($pickupLocation: [Float]!, $dropoffLocation: [Float]!, $description: String!) {
     addShipment(pickupLocation: $pickupLocation, dropoffLocation: $dropoffLocation, description: $description) {
       shipments {
         pickupLocation
